+++ src/services/scriptingservice.h
@@ -6,6 +6,7 @@
 #include <QtCore/QFile>
 #include <entities/script.h>
 #include <entities/notefolder.h>
+#include <QMimeData>
 
 
 struct ScriptComponent {
@@ -23,18 +24,22 @@ public:
     static ScriptingService *createInstance(QObject *parent);
     QQmlEngine* engine();
     void initComponents();
-    QString callModifyMediaMarkdown(QFile *file, QString markdownText);
+    QString callInsertMediaHook(QFile *file, QString markdownText);
     static bool validateScript(Script script, QString &errorMessage);
     Q_INVOKABLE bool startDetachedProcess(QString executablePath,
                                           QStringList parameters);
     Q_INVOKABLE QString currentNoteFolderPath();
+    QString callInsertingFromMimeDataHookForObject(QObject *object,
+                                                   const QMimeData *mimeData);
+    QString callInsertingFromMimeDataHook(const QMimeData *mimeData);
+
 private:
     QQmlEngine *_engine;
     QHash<int, ScriptComponent> _scriptComponents;
     bool methodExistsForObject(QObject *object, QString method);
-    QString callModifyMediaMarkdownForObject(QObject *object,
-                                             QFile *file,
-                                             QString markdownText);
+    QString callInsertMediaHookForObject(QObject *object,
+                                         QFile *file,
+                                         QString markdownText);
     void initComponent(Script script);
     void outputMethodsOfObject(QObject *object);